requirements.txt
# === 核心 ===
python-pptx>=1.0.0            # PPTX 生成
matplotlib>=3.8.0              # 统计图表渲染
seaborn>=0.13.0                # 统计图表（美观默认样式）
Pillow>=10.0.0                 # 图片处理
numpy>=1.24.0                  # 数值计算

# === 内容获取 ===
pypdf>=4.0.0                   # PDF 文本提取
python-docx>=1.1.0             # DOCX 解析

# === LLM ===
openai>=1.0.0                  # OpenAI 兼容 API (Kimi/Grok/DeepSeek/...)
python-dotenv>=1.0.0           # .env 环境变量
httpx>=0.27.0                  # HTTP 客户端
tenacity>=8.2.0                # 重试机制

# === 可选 ===
# pdfplumber>=0.11.0           # PDF 表格提取（增强）
# pdf2image>=1.17.0            # PDF → 图片（风格提取）
# pyppeteer>=1.0.2             # URL 爬取（首选）
# playwright>=1.40.0           # URL 爬取（备用）
# anthropic                    # Claude provider
# google-generativeai          # Gemini provider
# pyyaml>=6.0                  # Markdown YAML front matter 解析
